Refactor fork & push functionality
